% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demgeneration.R
\name{demsimulation}
\alias{demsimulation}
\title{Simulate DEMs based on a semi-variogram}
\usage{
demsimulation(target_raster, sv_model, maxdist = 0.01, nsim = 10)
}
\arguments{
\item{target_raster}{raster: The raster to make simulations from.This should
be projected. It is advised to set water bodies to NA using a water mask
(e.g Pekel et al., 2016). See Hawker et al., 2018 for guidance.}

\item{sv_model}{variogramModel: Select one from the pre-computed options
using modelselect or create your own}

\item{maxdist}{number(optional): from the gstat package to consider only
observations within a distance of maxdist from the target_raster location
are used for prediction or simulation. Default = 0.01}

\item{nsim}{number(optional): Number of simulations. Default = 10}
}
\value{
Simulated Rasters: In the form of a raster stack
}
\description{
\code{demsimulation} returns simulated DEMs
}
