% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demgeneration_LC.R
\name{demsimulation_LC}
\alias{demsimulation_LC}
\title{Simulate DEMs based on a semi-variogram based on landcover}
\usage{
demsimulation_LC(target_raster, LC_map, maxdist = 0.01, nsim = 10)
}
\arguments{
\item{target_raster}{raster: The raster to make simulations from.This should
be projected. It is advised to set water bodies to NA using a water mask
(e.g Pekel et al., 2016). See Hawker et al., 2018 for guidance.}

\item{LC_map}{Landcover map generated using download.CCI}

\item{maxdist}{number(optional): from the gstat package to consider only
observations within a distance of maxdist from the target_raster location
are used for prediction or simulation. Default = 0.01}

\item{nsim}{number(optional): Number of simulations. Default = 10}
}
\value{
Simulated Rasters: In the form of a raster stack
}
\description{
\code{demsimulation_LC} returns simulated DEMs
}
