% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.CCI.R
\name{download.CCI}
\alias{download.CCI}
\title{Download CCI Landcover data}
\usage{
download.CCI(targetDEM)
}
\arguments{
\item{targetDEM}{: DEM that will be simulated.}
}
\value{
Global CCI Landcover map
}
\description{
\code{download.CCI} Downloads CCI Landcover map into temporary directory
}
