% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRasterDEM.R
\name{writeRasterDEM}
\alias{writeRasterDEM}
\title{Save Simulated DEM}
\usage{
writeRasterDEM(DEMStack, filename = "", format = "GTiff")
}
\arguments{
\item{DEMStack}{rasterStack: Raster stack of simulated DEMs
generated using demgeneration}
}
\value{
Simulated Rasters: In the form of indidual rasters
}
\description{
\code{writeRasterDEM} returns simulated DEMs
}
