% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demsimulation_LC.R
\name{demsimulation_LC}
\alias{demsimulation_LC}
\title{Simulate DEMs based on a semi-variogram based on landcover}
\usage{
demsimulation_LC(DEM, LC_map, sv = "MERIT", maxdist = 0.01, nsim = 10,
  debuglevel = -1)
}
\arguments{
\item{DEM}{raster: The raster to make simulations from.This should
be projected. It is advised to set water bodies to NA using a water mask
(e.g Pekel et al., 2016). See Hawker et al., 2018 for guidance.}

\item{LC_map}{Landcover map generated using download.CCI}

\item{sv}{Semi-variogram set to use. Either 'MERIT' or 'SRTM'}

\item{maxdist}{number(optional): from the gstat package to consider only
observations within a distance of maxdist from the target_raster location
are used for prediction or simulation. Default = 0.01}

\item{nsim}{number(optional): Number of simulations. Default = 10}

\item{debuglevel}{(optional)  Useful values for debuglevel: 0: suppres any output except warning and error messages; 1:
normal output (default): short data report, program action and mode, program progress in %, total
execution time; 2: print the value of all global variables, all files read and written, and include
source file name and line number in error messages; 4: print OLS and WLS fit diagnostics; 8: print
all data after reading them; 16: print the neighbourhood selection for each prediction location; 32:
print (generalised) covariance matrices, design matrices, solutions, kriging weights, etc.; 64: print
variogram fit diagnostics (number of iterations and variogram model in each iteration step) and
order relation violations (indicator kriging values before and after order relation correction); 512:
print block (or area) discretization data for each prediction location. To combine settings, sum their
respective values. Negative values for debug.level are equal to positive, but cause the progress
counter to work}
}
\value{
Simulated Rasters: In the form of a raster stack
}
\description{
\code{demsimulation_LC} returns simulated DEMs
}
