% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselect.r
\name{modelselect}
\alias{modelselect}
\title{Read in semi-variogram for simulation}
\usage{
modelselect(model = "")
}
\arguments{
\item{model}{semi-variogram model to use in simulation
Choose from one of 'Burdekin', 'Ebro', 'Fens DSM', Fens DTM','Mekong','Mississippi','Po Delta'
or 'Wax lake'.}
}
\value{
Semi-variogram Model
}
\description{
\code{modelselect} Select Semi-Variogram for DEM Simulation
}
